// packages/dom/src/Visualizations/Enums/IndicatorVisualizationType.ts
export enum IndicatorVisualizationType {
    YearToDatePreviousYear = "YearToDatePreviousYear", //KpiTimePeriod
    QuarterToDatePreviousQuarter = "QuarterToDatePreviousQuarter", //KpiTimePeriod
    QuarterToDatePreviousYear = "QuarterToDatePreviousYear", //KpiTimePeriod
    MonthToDatePreviousMonth = "MonthToDatePreviousMonth", //KpiTimePeriod
    MonthToDatePreviousYear = "MonthToDatePreviousYear", //KpiTimePeriod
    LastYears = "LastYears", //SparklineAggregationType
    LastMonths = "LastMonths", //SparklineAggregationType
    LastDays = "LastDays", //SparklineAggregationType
}