// packages/dom/src/Visualizations/Enums/KpiTimePeriod.ts
export enum KpiTimePeriod {
    MonthToDatePreviousMonth = "MonthToDatePreviousMonth",
    MonthToDatePreviousYear = "MonthToDatePreviousYear",
    QuarterToDatePreviousQuarter = "QuarterToDatePreviousQuarter",
    QuarterToDatePreviousYear = "QuarterToDatePreviousYear",
    YearToDatePreviousYear = "YearToDatePreviousYear",
}